refactor(cl): convert PrayAndPayTestCase to a mixin
