+++ cl/favorites/tests.py
@@ -42,7 +42,6 @@ from cl.favorites.utils import (
 )
 from cl.lib.test_helpers import (
     AudioTestCase,
-    PrayAndPayTestCase,
     SimpleUserDataMixin,
 )
 from cl.search.factories import RECAPDocumentFactory
@@ -50,6 +49,7 @@ from cl.search.views import get_homepage_stats
 from cl.tests.base import SELENIUM_TIMEOUT, BaseSeleniumTest
 from cl.tests.cases import APITestCase, TestCase
 from cl.tests.fakes import FakeAvailableConfirmationPage, FakeConfirmationPage
+from cl.tests.mixins import PrayAndPayMixin
 from cl.tests.utils import make_client
 from cl.users.factories import UserFactory, UserProfileWithParentsFactory
 
@@ -663,7 +663,7 @@ class APITests(APITestCase, TestCase):
         self.assertEqual(response.json()["count"], 1)
 
 
-class RECAPPrayAndPay(SimpleUserDataMixin, PrayAndPayTestCase):
+class RECAPPrayAndPay(TestCase, SimpleUserDataMixin, PrayAndPayMixin):
     @override_settings(ALLOWED_PRAYER_COUNT=2)
     async def test_prayer_eligible(self) -> None:
         """Does the prayer_eligible method work properly?"""
@@ -1434,7 +1434,7 @@ class RECAPPrayAndPay(SimpleUserDataMixin, PrayAndPayTestCase):
 
 @patch("cl.favorites.utils.prayer_eligible", return_value=(True, 5))
 @patch("cl.favorites.signals.prayer_unavailable", wraps=prayer_unavailable)
-class PrayAndPaySignalTests(PrayAndPayTestCase):
+class PrayAndPaySignalTests(TestCase, PrayAndPayMixin):
     @patch("cl.favorites.signals.check_prayer_pacer")
     async def test_create_prayer_no_pacer_doc_id(
         self,
@@ -1561,7 +1561,7 @@ class PrayAndPaySignalTests(PrayAndPayTestCase):
 
 @patch("cl.favorites.tasks.get_or_cache_pacer_cookies")
 @patch("cl.favorites.tasks.prayer_unavailable", wraps=prayer_unavailable)
-class PrayAndPayCheckAvailabilityTaskTests(PrayAndPayTestCase):
+class PrayAndPayCheckAvailabilityTaskTests(TestCase, PrayAndPayMixin):
     @patch(
         "cl.favorites.tasks.DownloadConfirmationPage", new=FakeConfirmationPage
     )
@@ -1702,7 +1702,7 @@ class PrayAndPayCheckAvailabilityTaskTests(PrayAndPayTestCase):
         mock_check_prayer_pacer.delay.assert_called_once()
 
 
-class PrayerAPITests(PrayAndPayTestCase):
+class PrayerAPITests(TestCase, PrayAndPayMixin):
     """Check that Prayer API operations work as expected."""
 
     def setUp(self) -> None:
